fix filenames/ fix directory names/ refactor functions for button and input
