+++ my-app/src/components/hotelInfoSection/HotelItemMaker.jsx
