+++ my-app/src/components/hotelInfoSection/MainSection.jsx
