+++ my-app/src/components/hotelInfoSection/index.css
