+++ my-app/src/components/userSearch/index.css
@@ -16,6 +16,7 @@
     height: 64px;
     border: none;
     border-radius: 8px;
+    padding: 15px;
     outline: none;
 }
