+++ my-app/src/components/userSearch/searchInput/SearchInput.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const SearchInput = ({ setUserStr, text }) => (
+const SearchInput = ({ onChangeFunc, text }) => (
   <>
     <input
       className="form_input top_section__form_input--destination text_lg"
       placeholder={text}
-      onChange={(e) => setUserStr(e.target.value)}
+      onChange={(e) => onChangeFunc(e)}
     />
   </>
 );
 
 SearchInput.proTypes = {
   text: PropTypes.string.isRequired,
-  setUserStr: PropTypes.func.isRequired,
+  onChangeFunc: PropTypes.func.isRequired,
 };
 
 export default SearchInput;
