+++ my-app/src/helpersFunctions/textFormatter.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line no-unused-vars
 import React from 'react';
 
 const textFormatter = (x) => x.toLowerCase().trim();
